pyproject.toml
[project]
name = "vision-agents-mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "vision-agents[gemini,getstream]>=0.2.6",
]
